requirements.txt
django
django-environ
django-storages
google-cloud-storage
google-auth
Pillow
martor
psycopg2-binary
gunicorn
whitenoise
emojiflag
django-modeltranslation
django-photologue
